pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.2",
    "google-genai>=1.31.0",
    "python-telegram-bot>=22.3",
    "sift-stack-py>=0.8.4",
    "telegram>=0.0.1",
]
